Module encoding-index-singlebyte::windows_1257Stable

Windows code page 1257.

Functions

backward
Returns the index pointer for code point code in this index.

forward
Returns the index code point for pointer code in this index.